Refactor validation logic to decide comparison type in compilation and remove utility method for normalizing numbers

+++ javalidation-jakarta-validator-processor/src/test/java/test/jakarta/MaxReferenceRecord.java
@@ -4,4 +4,4 @@ import io.github.raniagus.javalidation.validator.*;
 import jakarta.validation.constraints.*;
 
 @Validate
-public record MaxReferenceRecord(@Max(100) Integer value) {}
+public record MaxReferenceRecord(@Max(100) Long value) {}
